requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.23
django-ckeditor==6.7.3
django-ckeditor-5==0.2.18
django-jazzmin==4.0.0
django-js-asset==2.2.0
pillow==10.4.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
